migrated inc button and labeled input
