+++ static/.gitignore
@@ -1,4 +1,1 @@
-*.d.ts
-*.js
-*.json
-*.wasm
+*
